Default commit_time_batch_for_recovery to true
Summary:
commit_time_batch_for_recovery is a performance config that
needs to be true to get high performance when write_policy is set to
write_prepared. So far teh default was FALSE but it makes the
configuration simpler if we set the default to TRUE. If needed for
debugging purposes, the debugger could disable the default value.

Differential Revision: D10394003