PS-7955 Temporary fix encrypt/decryt functions use in alter table
https://jira.percona.com/browse/PS-7949

***

`log_tmp_block_decrypt` uses the argument `offs` to work out the
initialization vector which was used in encryption. If the provided
`offs` is not the same as the one use in encryption the decrypted data
will be corrupted.

MySQL 8.0.27 has refactored the code so the `size` and `offs` are not
the same during write as during read so the current
`log_tmp_block_decrypt` is returning corrupted data.

Another upstream change, that affects these functions, is that the size
is not longer a multiple of MY_AES_BLOCK_SIZE. This causes we can't
neither use `my_aes_mode::ECB`.