PS-8047 - Adapt template reusing to prefix index queries optimization
https://jira.percona.com/browse/PS-8047

***

*Problem*:

mysql@589208e8656123511481a90d4967adb09a53661a introduces an optimization to reuse the prebuilt template. As part of this commit, in DEBUG mode, there is a check to verify the reused template is equal to a new created one.

This clashes with a Percona optimization in `row_search_mvcc` that modifies the template after its creation. More info on <https://www.percona.com/doc/percona-server/8.0/performance/prefix_index_queries_optimization.html>. This modification triggers the check above as in the following steps:

- `index_read` > `build_template`: creates template
- `row_search_mvcc`: Percona optimization modifies the template in order to get the values from the index.

    ```
        templ->rec_field_no = templ->rec_prefix_field_no;
    ```

- `index_read`: reuses template. In DEBUG, it checks that a new template is the same as the reused one. This check fails as the template as been modified.

We need to verify that reusing the template is safe. Different partitions might have different record sizes which might cause that, due to the Percona optimization, the value of the template `rec_field_no` varies for each one of the partition.

*Solution*:

We have verified that `row_search_mvcc` is not called by `build_template`. This means we can reuse the template. If it is reused it, the Percona optimization is going to be triggered anyway and it will adapt the value of `rec_field_no` for the current partition.

We have modified the MySQL check to obviate the field `rec_field_no`.