FB8-276 - Fix test tmp_table_bytes_written
***

Problem:

The `select ... group by ...` orders generate a tmp table due to the
`group by` order does not match the table order.

https://dev.mysql.com/doc/refman/8.0/en/group-by-optimization.html

Solution:

Create an index in the table so the rows are ordered by the same
criteria as the `group by` clause. This way no tmp table is generated.