PS-8047 - Fix test innodb.hp_deadlock
https://jira.percona.com/browse/PS-8047

Unlike MySQL mtr, PS mtr checks the variable `debug_sync` hasn't changed
after the test execution.

MySQL 8.0.29 sets the signal `con1_will_wait` but it doesn't clear it out.

The signal is not necessary as it is not used anywhere in the test.

PS needs to clear out that signal for the tests to pass.